[reversetunnel] agent side stale conn timeout
This commit introduces a watchdog mechanism to agent side ssh connections, allowing
the agent to detect stale connections and reconnect when needed.

A new ssh channel has been introduced: 'teleport-heartbeat-v2'. This channel is never
expected to be openned, instead it uses a private range reject reason to signal to the
agent that the server will reply to ping messages on this channel. It is used for feature
detection only and is only needed for backwards compatiblity during migration.

The agent mirrors the server side behaviour and calculates the watchdog to be:
`keepAlive * keepAliveCount`,

The agent behaviour can be disabled by setting `TELEPORT_UNSTABLE_DISABLE_AGENT_STALE_CONN_TIMEOUT` to `yes`
or other truthy value.

On older servers, the channel name is unknown and as such the agent will fallback to legacy behaviour until
the migration window is complete.

For legacy heartbeat behaviour the agent has always set the WantReply flag to false, so the server side code can remain unchanged.

Changelog: Adds support for reverse tunnel agent stale connection timeout, triggering a reconnect.